requirements.txt
Django>=5.2,<5.3
python-dotenv>=1.0.0
python-chess>=1.999